BwTreeIndex: Add visibility logic, abort and deferred delete logic, descending and limit scans (#356)
