+++ src/transaction/transaction_manager.cpp
@@ -189,6 +189,7 @@ TransactionQueue TransactionManager::CompletedTransactionsForGC() {
 }
 
 void TransactionManager::DeferAction(Action a) {
+  TERRIER_ASSERT(GCEnabled(), "Need GC enabled for deferred actions to be executed.");
   common::SpinLatch::ScopedSpinLatch guard(&deferred_actions_latch_);
   deferred_actions_.push({time_.load(), a});
 }
